README: Add list of personas
I often find myself going to my old PR with the example links just to
find a link to invoke the research/glossary persona. Create a list of
the personas in this repository with clickable links right in the
README.

This commit also appends a newline to the last line of the README, as
any sensible editor will do by default.